backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
requests==2.31.0
httpx==0.25.2
llama-index==0.10.15
openai==1.3.5
pinecone-client==3.2.2
tiktoken==0.5.2
langdetect==1.0.9
whisper==1.1.10
redis==5.0.1
celery==5.3.4
python-telegram-bot==20.7